requirements.txt
appdirs==1.4.4
astor==0.8.1
async-generator==1.10
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
cachetools==5.2.0
#certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fire==0.5.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
fonttools==4.38.0
future==0.18.2
h11==0.14.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lmdb==1.4.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
multiprocess==0.70.14
networkx==2.8.8
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.2
paddlepaddle==2.4.1
pandas==1.5.2
pdf2docx==0.5.6
Pillow==9.3.0
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.16.0
PyMuPDF==1.20.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
rapidfuzz==2.13.6
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
selenium==4.7.2
shapely==2.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
termcolor==2.1.1
tifffile==2022.10.10
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
visualdl==2.4.1
webdriver-manager==3.8.5
webdrivermanager==0.10.0
Werkzeug==2.2.2
wincertstore==0.2
wsproto==1.2.0
zipp==3.11.0
